+++ libraries/exoplayer_dash/src/main/java/androidx/media3/exoplayer/dash/DefaultDashChunkSource.java
@@ -63,6 +63,7 @@ import androidx.media3.exoplayer.upstream.LoaderErrorThrower;
 import androidx.media3.extractor.ChunkIndex;
 import androidx.media3.extractor.text.SubtitleParser;
 import com.google.common.collect.ImmutableMap;
+import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -113,18 +114,19 @@ public class DefaultDashChunkSource implements DashChunkSource {
       this.maxSegmentsPerLoad = maxSegmentsPerLoad;
     }
 
-    /**
-     * Sets the {@link SubtitleParser.Factory} to be used for parsing subtitles during extraction,
-     * or null to parse subtitles during decoding.
-     *
-     * <p>This may only be used with {@link BundledChunkExtractor.Factory}.
-     */
-    /* package */ Factory setSubtitleParserFactory(
-        @Nullable SubtitleParser.Factory subtitleParserFactory) {
-      if (chunkExtractorFactory instanceof BundledChunkExtractor.Factory) {
-        ((BundledChunkExtractor.Factory) chunkExtractorFactory)
-            .experimentalSetSubtitleParserFactory(subtitleParserFactory);
-      }
+    @CanIgnoreReturnValue
+    @Override
+    public Factory setSubtitleParserFactory(SubtitleParser.Factory subtitleParserFactory) {
+      chunkExtractorFactory.setSubtitleParserFactory(subtitleParserFactory);
+      return this;
+    }
+
+    @CanIgnoreReturnValue
+    @Override
+    public Factory experimentalParseSubtitlesDuringExtraction(
+        boolean parseSubtitlesDuringExtraction) {
+      chunkExtractorFactory.experimentalParseSubtitlesDuringExtraction(
+          parseSubtitlesDuringExtraction);
       return this;
     }