+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/DefaultHlsExtractorFactory.java
@@ -33,6 +33,7 @@ import androidx.media3.extractor.Extractor;
 import androidx.media3.extractor.ExtractorInput;
 import androidx.media3.extractor.mp3.Mp3Extractor;
 import androidx.media3.extractor.mp4.FragmentedMp4Extractor;
+import androidx.media3.extractor.text.DefaultSubtitleParserFactory;
 import androidx.media3.extractor.text.SubtitleParser;
 import androidx.media3.extractor.ts.Ac3Extractor;
 import androidx.media3.extractor.ts.Ac4Extractor;
@@ -41,6 +42,7 @@ import androidx.media3.extractor.ts.DefaultTsPayloadReaderFactory;
 import androidx.media3.extractor.ts.TsExtractor;
 import com.google.common.collect.ImmutableList;
 import com.google.common.primitives.Ints;
+import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.io.EOFException;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -67,8 +69,8 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
 
   private final @DefaultTsPayloadReaderFactory.Flags int payloadReaderFactoryFlags;
 
-  /** Non-null if subtitles should be parsed during extraction, null otherwise. */
-  @Nullable private SubtitleParser.Factory subtitleParserFactory;
+  private SubtitleParser.Factory subtitleParserFactory;
+  private boolean parseSubtitlesDuringExtraction;
 
   private final boolean exposeCea608WhenMissingDeclarations;
 
@@ -96,6 +98,7 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
       int payloadReaderFactoryFlags, boolean exposeCea608WhenMissingDeclarations) {
     this.payloadReaderFactoryFlags = payloadReaderFactoryFlags;
     this.exposeCea608WhenMissingDeclarations = exposeCea608WhenMissingDeclarations;
+    subtitleParserFactory = new DefaultSubtitleParserFactory();
   }
 
   @Override
@@ -135,7 +138,11 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
               createExtractorByFileType(fileType, format, muxedCaptionFormats, timestampAdjuster));
       if (sniffQuietly(extractor, sniffingExtractorInput)) {
         return new BundledHlsMediaChunkExtractor(
-            extractor, format, timestampAdjuster, subtitleParserFactory);
+            extractor,
+            format,
+            timestampAdjuster,
+            subtitleParserFactory,
+            parseSubtitlesDuringExtraction);
       }
       if (fallBackExtractor == null
           && (fileType == formatInferredFileType
@@ -149,23 +156,26 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
     }
 
     return new BundledHlsMediaChunkExtractor(
-        checkNotNull(fallBackExtractor), format, timestampAdjuster, subtitleParserFactory);
+        checkNotNull(fallBackExtractor),
+        format,
+        timestampAdjuster,
+        subtitleParserFactory,
+        parseSubtitlesDuringExtraction);
   }
 
-  /**
-   * Sets the {@link SubtitleParser.Factory} to use for parsing subtitles during extraction, or null
-   * to parse subtitles during decoding. The default is null (subtitles parsed after decoding).
-   *
-   * <p>This method is experimental. Its default value may change, or it may be renamed or removed
-   * in a future release.
-   *
-   * @param subtitleParserFactory The {@link SubtitleParser.Factory} for parsing subtitles during
-   *     extraction.
-   * @return This factory, for convenience.
-   */
-  public DefaultHlsExtractorFactory experimentalSetSubtitleParserFactory(
-      @Nullable SubtitleParser.Factory subtitleParserFactory) {
+  @CanIgnoreReturnValue
+  @Override
+  public DefaultHlsExtractorFactory setSubtitleParserFactory(
+      SubtitleParser.Factory subtitleParserFactory) {
     this.subtitleParserFactory = subtitleParserFactory;
+    return this;
+  }
+
+  @CanIgnoreReturnValue
+  @Override
+  public DefaultHlsExtractorFactory experimentalParseSubtitlesDuringExtraction(
+      boolean parseSubtitlesDuringExtraction) {
+    this.parseSubtitlesDuringExtraction = parseSubtitlesDuringExtraction;
     return this;
   }
 
@@ -176,18 +186,15 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
    * MimeTypes#APPLICATION_MEDIA3_CUES} if it is supported by {@link SubtitleParser.Factory}.
    *
    * <p>To modify the support behavior, you can {@linkplain
-   * #experimentalSetSubtitleParserFactory(SubtitleParser.Factory) set your own subtitle parser
-   * factory}.
+   * #setSubtitleParserFactory(SubtitleParser.Factory) set your own subtitle parser factory}.
    */
   @Override
   public Format getOutputTextFormat(Format sourceFormat) {
-    if (subtitleParserFactory != null && subtitleParserFactory.supportsFormat(sourceFormat)) {
-      @Format.CueReplacementBehavior
-      int cueReplacementBehavior = subtitleParserFactory.getCueReplacementBehavior(sourceFormat);
+    if (parseSubtitlesDuringExtraction && subtitleParserFactory.supportsFormat(sourceFormat)) {
       return sourceFormat
           .buildUpon()
           .setSampleMimeType(MimeTypes.APPLICATION_MEDIA3_CUES)
-          .setCueReplacementBehavior(cueReplacementBehavior)
+          .setCueReplacementBehavior(subtitleParserFactory.getCueReplacementBehavior(sourceFormat))
           .setCodecs(
               sourceFormat.sampleMimeType
                   + (sourceFormat.codecs != null ? " " + sourceFormat.codecs : ""))
@@ -216,16 +223,10 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
     // LINT.IfChange(extractor_instantiation)
     switch (fileType) {
       case FileTypes.WEBVTT:
-        SubtitleParser.Factory webvttSubtitleParserFactory = SubtitleParser.Factory.UNSUPPORTED;
-        boolean parseSubtitlesDuringExtraction = false;
-        if (subtitleParserFactory != null && subtitleParserFactory.supportsFormat(format)) {
-          webvttSubtitleParserFactory = subtitleParserFactory;
-          parseSubtitlesDuringExtraction = true;
-        }
         return new WebvttExtractor(
             format.language,
             timestampAdjuster,
-            webvttSubtitleParserFactory,
+            subtitleParserFactory,
             parseSubtitlesDuringExtraction);
       case FileTypes.ADTS:
         return new AdtsExtractor();
@@ -237,7 +238,11 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
         return new Mp3Extractor(/* flags= */ 0, /* forcedFirstSampleTimestampUs= */ 0);
       case FileTypes.MP4:
         return createFragmentedMp4Extractor(
-            subtitleParserFactory, timestampAdjuster, format, muxedCaptionFormats);
+            subtitleParserFactory,
+            parseSubtitlesDuringExtraction,
+            timestampAdjuster,
+            format,
+            muxedCaptionFormats);
       case FileTypes.TS:
         return createTsExtractor(
             payloadReaderFactoryFlags,
@@ -245,7 +250,8 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
             format,
             muxedCaptionFormats,
             timestampAdjuster,
-            subtitleParserFactory);
+            subtitleParserFactory,
+            parseSubtitlesDuringExtraction);
       default:
         return null;
     }
@@ -257,7 +263,8 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
       Format format,
       @Nullable List<Format> muxedCaptionFormats,
       TimestampAdjuster timestampAdjuster,
-      @Nullable SubtitleParser.Factory subtitleParserFactory) {
+      SubtitleParser.Factory subtitleParserFactory,
+      boolean parseSubtitlesDuringExtraction) {
     @DefaultTsPayloadReaderFactory.Flags
     int payloadReaderFactoryFlags =
         DefaultTsPayloadReaderFactory.FLAG_IGNORE_SPLICE_INFO_STREAM
@@ -287,7 +294,7 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
       }
     }
     @TsExtractor.Flags int extractorFlags = 0;
-    if (subtitleParserFactory == null) {
+    if (!parseSubtitlesDuringExtraction) {
       subtitleParserFactory = SubtitleParser.Factory.UNSUPPORTED;
       extractorFlags |= TsExtractor.FLAG_EMIT_RAW_SUBTITLE_DATA;
     }
@@ -301,7 +308,8 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
   }
 
   private static FragmentedMp4Extractor createFragmentedMp4Extractor(
-      @Nullable SubtitleParser.Factory subtitleParserFactory,
+      SubtitleParser.Factory subtitleParserFactory,
+      boolean parseSubtitlesDuringExtraction,
       TimestampAdjuster timestampAdjuster,
       Format format,
       @Nullable List<Format> muxedCaptionFormats) {
@@ -309,7 +317,7 @@ public final class DefaultHlsExtractorFactory implements HlsExtractorFactory {
     // creating a separate EMSG track for every audio track in a video stream.
     @FragmentedMp4Extractor.Flags
     int flags = isFmp4Variant(format) ? FragmentedMp4Extractor.FLAG_ENABLE_EMSG_TRACK : 0;
-    if (subtitleParserFactory == null) {
+    if (!parseSubtitlesDuringExtraction) {
       subtitleParserFactory = SubtitleParser.Factory.UNSUPPORTED;
       flags |= FragmentedMp4Extractor.FLAG_EMIT_RAW_SUBTITLE_DATA;
     }