+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsExtractorFactory.java
@@ -26,6 +26,8 @@ import androidx.media3.exoplayer.analytics.PlayerId;
 import androidx.media3.extractor.Extractor;
 import androidx.media3.extractor.ExtractorInput;
 import androidx.media3.extractor.PositionHolder;
+import androidx.media3.extractor.text.SubtitleParser;
+import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -62,6 +64,37 @@ public interface HlsExtractorFactory {
       ExtractorInput sniffingExtractorInput,
       PlayerId playerId)
       throws IOException;
+
+  /**
+   * Sets the {@link SubtitleParser.Factory} to use for parsing subtitles during extraction. The
+   * default factory value is implementation dependent.
+   *
+   * @param subtitleParserFactory The {@link SubtitleParser.Factory} for parsing subtitles during
+   *     extraction.
+   * @return This factory, for convenience.
+   */
+  @CanIgnoreReturnValue
+  default HlsExtractorFactory setSubtitleParserFactory(
+      SubtitleParser.Factory subtitleParserFactory) {
+    return this;
+  }
+
+  /**
+   * Sets whether subtitles should be parsed as part of extraction (before being added to the sample
+   * queue) or as part of rendering (when being taken from the sample queue). Defaults to {@code
+   * false} (i.e. subtitles will be parsed as part of rendering).
+   *
+   * <p>This method is experimental and will be renamed or removed in a future release.
+   *
+   * @param parseSubtitlesDuringExtraction Whether to parse subtitles during extraction or
+   *     rendering.
+   * @return This factory, for convenience.
+   */
+  @CanIgnoreReturnValue
+  default HlsExtractorFactory experimentalParseSubtitlesDuringExtraction(
+      boolean parseSubtitlesDuringExtraction) {
+    return this;
+  }
 
   /**
    * Returns the output {@link Format} of emitted {@linkplain C#TRACK_TYPE_TEXT text samples} which