+++ RELEASENOTES.md
@@ -15,6 +15,11 @@
         implementing a custom `CompositeSequenceableLoaderFactory`.
     *   Fix issue where repeating the same time causes metadata from this item
         to be cleared ([#1007](https://github.com/androidx/media/issues/1007)).
+    *   Rename `experimentalSetSubtitleParserFactory` methods on
+        `BundledChunkExtractor.Factory` and `DefaultHlsExtractorFactory` to
+        `setSubtitleParserFactory` and disallow passing `null`. Use the new
+        `experimentalParseSubtitlesDuringExtraction(boolean)` methods to control
+        parsing behaviour.
 *   Transformer:
 *   Track Selection:
 *   Extractors: